+++ docs/pages/_en/1.0/reference/release-notes.md
@@ -13,6 +13,7 @@ group: Deployment & Configs
 ### RC
 
 - Added 'Appearance Asset Identity' component.
+- Renamed 'Add Linear Dimension' component to 'Add Aligned Dimension'.
 
 {% endcapture %}